+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### [31.05.26 - 0.0.3]
+- removed pipeline annotation
+- fixed workflow for gitlab-ci and pipeline
+
 ### [31.05.26 - 0.0.2]
 - added JobbuilderImpl and PipelineBuilderImpl
 - updated CI schema